perf(ssr-runtime)!: sacar react-dom/server del bundle del navegador
`react-dom` no declara `sideEffects: false`, asi que Rollup no podia sacudirlo:
el import estatico de `react-dom/server` en el entry principal arrastraba la
libreria entera a cualquier chunk que tocara `@calumet/suamox`, y el router la
toca siempre. Ademas iba con `<link rel="modulepreload">`, o sea que se
descargaba y parseaba en cada visita para no ejecutarse nunca.

`renderPage` era la unica pieza que la usaba. Se mueve a su propio entry point,
`@calumet/suamox/server`, junto al helper que renderiza el head. El adaptador la
carga por el mismo runner de Vite que el resto del runtime, que es lo que
comparte la instancia de los contextos de React con las paginas.

Medido sobre examples/basic:

  antes    453.667 bytes de JS de cliente
  despues  256.429 bytes        -43%

El chunk de 197.238 bytes con renderToStaticMarkup, renderToString y
renderToReadableStream desaparece, y ya no esta en la lista de precarga.

BREAKING CHANGE: `renderPage` se importa de `@calumet/suamox/server`. Solo
afecta a codigo de servidor; nada del navegador la usaba.

Cierra #30.
