+++ packages/create-app/package.json
@@ -1,6 +1,6 @@
 {
   "name": "@calumet/suamox-create-app",
-  "version": "0.3.0",
+  "version": "0.3.1",
   "description": "Scaffold a new Suamox project",
   "type": "module",
   "bin": {
